requirements.txt
wheel
gunicorn
flask
pillow
displayhatmini
RPi.GPIO
spidev
